Abstract Animal from Play
Make a new Animal class and move all the Prey logic down into it.

I didn't just rename Prey, as that would break some of the editor links.

Animal is still not generic, but that's the next patch.  There was no modification to the Prey logic when moving to Animal in this patch.