Work from RDE #30 and added EvolvableStats
Generalized traits a little so we can evolve more than one.  In this example added evolving of the wanderSpeedMultiplier, since as far as I can tell wanderSpeedMultipler != 0 is an evolutionary disadvantage in this current setup.

This also contains the work from RDE #30 which was mostly tuning around having a single speed variable instad of multiple.